Make it all work, we can use echo example now.
