+++ rechannel/Cargo.toml
@@ -12,9 +12,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
 log = "0.4.17"
-serde = { version = "1.0", features = [ "derive" ] } 
-bincode = "1.3.1"
-bytes = { version = "1.1", features = ["serde"] }
+bytes = "1.1"
 octets = "0.2"
 
 [dev-dependencies]
