+++ rechannel/src/channels/unreliable.rs
@@ -5,13 +5,11 @@ use std::{
 
 use bytes::Bytes;
 
-use crate::{
-    another_channel::{Packet, Slice},
-    error::ChannelError,
-};
+use crate::{error::ChannelError, packet::Packet};
 
-use super::{SliceConstructor, SLICE_SIZE};
+use super::{slice_constructor::Slice, SliceConstructor, SLICE_SIZE};
 
+#[derive(Debug)]
 pub struct SendChannelUnreliable {
     channel_id: u8,
     unreliable_messages: VecDeque<Bytes>,
@@ -21,7 +19,9 @@ pub struct SendChannelUnreliable {
     error: Option<ChannelError>,
 }
 
+#[derive(Debug)]
 pub struct ReceiveChannelUnreliable {
+    channel_id: u8,
     messages: VecDeque<Bytes>,
     slices: HashMap<u64, SliceConstructor>,
     slices_last_received: HashMap<u64, Duration>,
@@ -31,7 +31,7 @@ pub struct ReceiveChannelUnreliable {
 }
 
 impl SendChannelUnreliable {
-    fn new(channel_id: u8, max_memory_usage_bytes: usize) -> Self {
+    pub fn new(channel_id: u8, max_memory_usage_bytes: usize) -> Self {
         Self {
             channel_id,
             unreliable_messages: VecDeque::new(),
@@ -42,14 +42,14 @@ impl SendChannelUnreliable {
         }
     }
 
-    fn get_messages_to_send(&mut self, packet_sequence: &mut u64) -> Vec<Packet> {
+    pub fn get_messages_to_send(&mut self) -> Vec<Packet> {
         let mut packets: Vec<Packet> = vec![];
         let mut small_messages: Vec<Bytes> = vec![];
         let mut small_messages_bytes = 0;
 
         while let Some(message) = self.unreliable_messages.pop_front() {
             if message.len() > SLICE_SIZE {
-                let num_slices = message.len() / SLICE_SIZE;
+                let num_slices = (message.len() + SLICE_SIZE - 1) / SLICE_SIZE;
 
                 for slice_index in 0..num_slices {
                     let start = slice_index * SLICE_SIZE;
@@ -63,25 +63,20 @@ impl SendChannelUnreliable {
                         payload,
                     };
 
-                    packets.push(Packet::MessageSlice {
-                        packet_sequence: *packet_sequence,
+                    packets.push(Packet::UnreliableSlice {
                         channel_id: self.channel_id,
                         slice,
                     });
-
-                    *packet_sequence += 1;
                 }
 
                 self.sliced_message_id += 1;
             } else {
                 if small_messages_bytes + message.len() > SLICE_SIZE {
                     packets.push(Packet::SmallUnreliable {
-                        packet_sequence: *packet_sequence,
                         channel_id: self.channel_id,
                         messages: std::mem::take(&mut small_messages),
                     });
                     small_messages_bytes = 0;
-                    *packet_sequence += 1;
                 }
 
                 small_messages_bytes += message.len();
@@ -92,17 +87,15 @@ impl SendChannelUnreliable {
         // Generate final packet for remaining small messages
         if !small_messages.is_empty() {
             packets.push(Packet::SmallUnreliable {
-                packet_sequence: *packet_sequence,
                 channel_id: self.channel_id,
                 messages: std::mem::take(&mut small_messages),
             });
-            *packet_sequence += 1;
         }
 
         packets
     }
 
-    fn send_message(&mut self, message: Bytes) {
+    pub fn send_message(&mut self, message: Bytes) {
         if self.max_memory_usage_bytes < self.memory_usage_bytes + message.len() {
             // TODO: log::warm
             return;
@@ -114,8 +107,9 @@ impl SendChannelUnreliable {
 }
 
 impl ReceiveChannelUnreliable {
-    fn new(max_memory_usage_bytes: usize) -> Self {
+    pub fn new(channel_id: u8, max_memory_usage_bytes: usize) -> Self {
         Self {
+            channel_id,
             slices: HashMap::new(),
             slices_last_received: HashMap::new(),
             messages: VecDeque::new(),
@@ -157,6 +151,7 @@ impl ReceiveChannelUnreliable {
                 self.slices.remove(&slice.message_id);
                 self.slices_last_received.remove(&slice.message_id);
                 self.memory_usage_bytes -= slice.num_slices * SLICE_SIZE;
+                self.memory_usage_bytes += message.len();
                 self.messages.push_back(message);
             }
             Ok(None) => {
@@ -166,20 +161,22 @@ impl ReceiveChannelUnreliable {
     }
 
     pub fn receive_message(&mut self) -> Option<Bytes> {
-        self.messages.pop_front()
+        let Some(message) = self.messages.pop_front() else {
+            return None
+        };
+        self.memory_usage_bytes -= message.len();
+        Some(message)
     }
 }
 
+#[cfg(test)]
 mod tests {
-    use crate::packet;
-
     use super::*;
 
     #[test]
     fn small_packet() {
         let max_memory: usize = 10000;
-        let mut sequence: u64 = 0;
-        let mut recv = ReceiveChannelUnreliable::new(max_memory);
+        let mut recv = ReceiveChannelUnreliable::new(0, max_memory);
         let mut send = SendChannelUnreliable::new(0, max_memory);
 
         let message1 = vec![1, 2, 3];
@@ -188,9 +185,9 @@ mod tests {
         send.send_message(message1.clone().into());
         send.send_message(message2.clone().into());
 
-        let packets = send.get_messages_to_send(&mut sequence);
+        let packets = send.get_messages_to_send();
         for packet in packets {
-            let Packet::SmallUnreliable { packet_sequence: 0, channel_id: 0, messages } = packet else {
+            let Packet::SmallUnreliable { channel_id: 0, messages } = packet else {
                 unreachable!();
             };
             for message in messages {
@@ -205,25 +202,24 @@ mod tests {
         assert_eq!(message1, new_message1);
         assert_eq!(message2, new_message2);
 
-        let packets = send.get_messages_to_send(&mut sequence);
+        let packets = send.get_messages_to_send();
         assert!(packets.is_empty());
     }
 
     #[test]
     fn slice_packet() {
         let max_memory: usize = 10000;
-        let mut sequence: u64 = 0;
         let current_time = Duration::ZERO;
-        let mut recv = ReceiveChannelUnreliable::new(max_memory);
+        let mut recv = ReceiveChannelUnreliable::new(0, max_memory);
         let mut send = SendChannelUnreliable::new(0, max_memory);
 
         let message = vec![5; SLICE_SIZE * 3];
 
         send.send_message(message.clone().into());
 
-        let packets = send.get_messages_to_send(&mut sequence);
+        let packets = send.get_messages_to_send();
         for packet in packets {
-            let Packet::MessageSlice { channel_id: 0, slice, .. } = packet else {
+            let Packet::UnreliableSlice { channel_id: 0, slice } = packet else {
                 unreachable!();
             };
             recv.process_slice(slice, current_time);
@@ -234,14 +230,13 @@ mod tests {
 
         assert_eq!(message, new_message);
 
-        let packets = send.get_messages_to_send(&mut sequence);
+        let packets = send.get_messages_to_send();
         assert!(packets.is_empty());
     }
 
     #[test]
     fn max_memory() {
-        let mut sequence: u64 = 0;
-        let mut recv = ReceiveChannelUnreliable::new(50);
+        let mut recv = ReceiveChannelUnreliable::new(0, 50);
         let mut send = SendChannelUnreliable::new(0, 40);
 
         let message = vec![5; 50];
@@ -249,9 +244,9 @@ mod tests {
         send.send_message(message.clone().into());
         send.send_message(message.clone().into());
 
-        let packets = send.get_messages_to_send(&mut sequence);
+        let packets = send.get_messages_to_send();
         for packet in packets {
-            let Packet::SmallUnreliable { packet_sequence: 0, channel_id: 0, messages } = packet else {
+            let Packet::SmallUnreliable { channel_id: 0, messages } = packet else {
                 unreachable!();
             };
