+++ rechannel/src/packet.rs
@@ -1,61 +1,402 @@
-use crate::error::DisconnectionReason;
+use bytes::Bytes;
+use std::ops::Range;
 
-use bincode::Options;
-use serde::{Deserialize, Serialize};
+use crate::channels::slice_constructor::Slice;
 
 pub type Payload = Vec<u8>;
 
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ChannelPacketData {
-    pub messages: Vec<Payload>,
-    pub channel_id: u8,
+#[derive(Debug, PartialEq, Eq)]
+pub enum Packet {
+    SmallReliable {
+        packet_sequence: u64,
+        channel_id: u8,
+        messages: Vec<(u64, Bytes)>,
+    },
+    SmallUnreliable {
+        channel_id: u8,
+        messages: Vec<Bytes>,
+    },
+    UnreliableSlice {
+        channel_id: u8,
+        slice: Slice,
+    },
+    ReliableSlice {
+        packet_sequence: u64,
+        channel_id: u8,
+        slice: Slice,
+    },
+    Ack {
+        packet_sequence: u64,
+        ack_ranges: Vec<Range<u64>>,
+    },
+    RequestAck,
+    Disconnect,
 }
 
-#[derive(Copy, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct AckData {
-    pub ack: u16,
-    pub ack_bits: u32,
-}
+impl Packet {
+    pub fn is_ack_eliciting(&self) -> bool {
+        match self {
+            Packet::SmallReliable { .. } | Packet::ReliableSlice { .. } | Packet::RequestAck => true,
+            _ => false,
+        }
+    }
+    pub fn to_bytes(&self, b: &mut octets::OctetsMut) -> Result<usize, octets::BufferTooShortError> {
+        let before = b.cap();
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub(crate) enum Packet {
-    Normal {
-        sequence: u16,
-        ack_data: AckData,
-        channels_packet_data: Vec<ChannelPacketData>,
-    },
-    Fragment {
-        sequence: u16,
-        ack_data: AckData,
-        fragment_data: FragmentData,
-    },
-    Heartbeat {
-        ack_data: AckData,
-    },
-    Disconnect {
-        reason: DisconnectionReason,
-    },
-}
+        match self {
+            Packet::SmallReliable {
+                packet_sequence,
+                channel_id,
+                messages,
+            } => {
+                b.put_u8(0)?;
+                b.put_u8(*channel_id)?;
+                b.put_varint(*packet_sequence)?;
+                b.put_u16(messages.len() as u16)?;
+                for (message_id, message) in messages {
+                    b.put_varint(*message_id)?;
+                    b.put_varint(message.len() as u64)?;
+                    b.put_bytes(message)?;
+                }
+            }
+            Packet::SmallUnreliable { channel_id, messages } => {
+                b.put_u8(1)?;
+                b.put_u8(*channel_id)?;
+                b.put_u16(messages.len() as u16)?;
+                for message in messages {
+                    b.put_varint(message.len() as u64)?;
+                    b.put_bytes(message)?;
+                }
+            }
+            Packet::ReliableSlice {
+                packet_sequence,
+                channel_id,
+                slice,
+            } => {
+                b.put_u8(2)?;
+                b.put_u8(*channel_id)?;
+                b.put_varint(*packet_sequence)?;
+                b.put_varint(slice.message_id)?;
+                b.put_varint(slice.slice_index as u64)?;
+                b.put_varint(slice.num_slices as u64)?;
+                b.put_varint(slice.payload.len() as u64)?;
+                b.put_bytes(&slice.payload)?;
+            }
+            Packet::UnreliableSlice { channel_id, slice } => {
+                b.put_u8(3)?;
+                b.put_u8(*channel_id)?;
+                b.put_varint(slice.message_id)?;
+                b.put_varint(slice.slice_index as u64)?;
+                b.put_varint(slice.num_slices as u64)?;
+                b.put_varint(slice.payload.len() as u64)?;
+                b.put_bytes(&slice.payload)?;
+            }
+            Packet::Ack {
+                packet_sequence,
+                ack_ranges,
+            } => {
+                b.put_u8(4)?;
+                b.put_varint(*packet_sequence)?;
 
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub(crate) struct FragmentData {
-    pub fragment_id: u8,
-    pub num_fragments: u8,
-    pub payload: Payload,
-}
+                // Consider this ranges:
+                // [20010..20020   ,  20035..20040]
+                //  <----10----><-15-><----5------>
+                //
+                // We can represented more compactly each range if we serialize it based
+                // on the start of the previous one, since the difference is usually small
+                // The ranges would become before serializing:
+                // 20040 5 1 15 10
+                //   |   | |  |  |
+                //   |   | |  |  +-> 10: size of 20017..20020
+                //   |   | |  +----> 15: gap between ranges 20010..20020 and 20035..20040
+                //   |   | +--------> 1: remaing number of ranges
+                //   |   +----------> 5: size of 20035..20040
+                //   +----------> 20040:  end of 20035..20040
+                //
+                // We can always reconstruct the ranges using the start of the previous one and the gap.
 
-impl std::fmt::Debug for AckData {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Ack")
-            .field("last ack", &self.ack)
-            .field("ack mask", &format!("{:031b}", &self.ack_bits))
-            .finish()
+                // Iterate in reverse order
+                let mut it = ack_ranges.iter().rev();
+
+                // Extract the last range (first in the iterator)
+                let last = it.next().unwrap();
+                let last_range_size = (last.end - 1) - last.start;
+
+                b.put_varint(last.end - 1)?;
+                b.put_varint(last_range_size)?;
+
+                // Write the number of remaining ranges
+                b.put_varint(it.len() as u64)?;
+
+                let mut previous_range_start = last.start;
+                // For each subsequent range:
+                for range in it {
+                    // Calculate the gap between the start of the previous range and the end of the current range
+                    let gap = previous_range_start - range.end - 1;
+                    let range_size = (range.end - 1) - range.start;
+
+                    b.put_varint(gap)?;
+                    b.put_varint(range_size)?;
+
+                    previous_range_start = range.start;
+                }
+            }
+            Packet::RequestAck => {
+                b.put_u8(5)?;
+            }
+            Packet::Disconnect => {
+                b.put_u8(6)?;
+            }
+        }
+
+        Ok(before - b.cap())
+    }
+
+    pub fn from_bytes(b: &mut octets::Octets) -> Result<Packet, octets::BufferTooShortError> {
+        let packet_type = b.get_u8()?;
+        match packet_type {
+            0 => {
+                // SmallReliable
+                let channel_id = b.get_u8()?;
+                let packet_sequence = b.get_varint()?;
+                let messages_len = b.get_u16()?;
+                let mut messages: Vec<(u64, Bytes)> = Vec::with_capacity(64);
+                for _ in 0..messages_len {
+                    let message_id = b.get_varint()?;
+                    let payload = b.get_bytes_with_varint_length()?;
+
+                    messages.push((message_id, payload.to_vec().into()));
+                }
+
+                Ok(Packet::SmallReliable {
+                    packet_sequence,
+                    channel_id,
+                    messages,
+                })
+            }
+            1 => {
+                // SmallUnreliable
+                let channel_id = b.get_u8()?;
+                let messages_len = b.get_u16()?;
+                let mut messages: Vec<Bytes> = Vec::with_capacity(64);
+                for _ in 0..messages_len {
+                    let payload = b.get_bytes_with_varint_length()?;
+                    messages.push(payload.to_vec().into());
+                }
+
+                Ok(Packet::SmallUnreliable { channel_id, messages })
+            }
+            2 => {
+                // ReliableSlice
+                let channel_id = b.get_u8()?;
+                let packet_sequence = b.get_varint()?;
+                let message_id = b.get_varint()?;
+                let slice_index = b.get_varint()? as usize;
+                let num_slices = b.get_varint()? as usize;
+                let payload = b.get_bytes_with_varint_length()?;
+
+                let slice = Slice {
+                    message_id,
+                    slice_index,
+                    num_slices,
+                    payload: payload.to_vec().into(),
+                };
+                Ok(Packet::ReliableSlice {
+                    packet_sequence,
+                    channel_id,
+                    slice,
+                })
+            }
+            3 => {
+                // UnreliableSlice
+                let channel_id = b.get_u8()?;
+                let message_id = b.get_varint()?;
+                let slice_index = b.get_varint()? as usize;
+                let num_slices = b.get_varint()? as usize;
+                let payload = b.get_bytes_with_varint_length()?;
+
+                let slice = Slice {
+                    message_id,
+                    slice_index,
+                    num_slices,
+                    payload: payload.to_vec().into(),
+                };
+                Ok(Packet::UnreliableSlice { channel_id, slice })
+            }
+            4 => {
+                // Ack
+                let packet_sequence = b.get_varint()?;
+
+                let first_range_end = b.get_varint()?;
+                let first_range_size = b.get_varint()?;
+                let num_remaining_ranges = b.get_varint()?;
+
+                if first_range_end < first_range_size {
+                    // TODO: Invalid ack packet
+                    return Err(octets::BufferTooShortError);
+                }
+
+                let mut ranges: Vec<Range<u64>> = Vec::with_capacity(32);
+
+                let first_range_start = first_range_end - first_range_size;
+                ranges.push(first_range_start..first_range_end + 1);
+
+                let mut previous_range_start = first_range_start;
+                for _ in 0..num_remaining_ranges {
+                    // Get the gap between the previous range and the current one
+                    let gap = b.get_varint()?;
+
+                    if previous_range_start < 2 + gap {
+                        // TODO: Invalid ack packet
+                        return Err(octets::BufferTooShortError);
+                    }
+
+                    // Get the end of the current range using the start of the previous one and the gap
+                    let range_end = (previous_range_start - gap) - 2;
+                    let range_size = b.get_varint()?;
+
+                    if range_end < range_size {
+                        // TODO: Invalid ack packet
+                        return Err(octets::BufferTooShortError);
+                    }
+
+                    let range_start = range_end - range_size;
+                    ranges.push(range_start..range_end + 1);
+
+                    previous_range_start = range_start;
+                }
+
+                ranges.reverse();
+
+                Ok(Packet::Ack {
+                    packet_sequence,
+                    ack_ranges: ranges,
+                })
+            }
+            5 => {
+                // Request Ack
+                Ok(Packet::RequestAck)
+            }
+            6 => {
+                // Disconnect
+                Ok(Packet::Disconnect)
+            }
+            _ => Err(octets::BufferTooShortError), // TODO: correct error (invalid packet type)
+        }
     }
 }
 
-/// Given a disconnect reason, serialize a disconnect packet to be sent.
-pub fn disconnect_packet(reason: DisconnectionReason) -> Result<Payload, bincode::Error> {
-    let packet = Packet::Disconnect { reason };
-    let packet = bincode::options().serialize(&packet)?;
-    Ok(packet)
+#[cfg(test)]
+mod tests {
+    use crate::channels::SLICE_SIZE;
+
+    use super::*;
+
+    #[test]
+    fn serialize_small_reliable_packet() {
+        let mut buffer = [0u8; 1300];
+        let packet = Packet::SmallReliable {
+            packet_sequence: 0,
+            channel_id: 0,
+            messages: vec![(0, vec![0, 0, 0].into()), (1, vec![1, 1, 1].into()), (2, vec![2, 2, 2].into())],
+        };
+
+        let mut b = octets::OctetsMut::with_slice(&mut buffer);
+        packet.to_bytes(&mut b).unwrap();
+
+        let mut b = octets::Octets::with_slice(&buffer);
+        let recv_packet = Packet::from_bytes(&mut b).unwrap();
+        assert_eq!(packet, recv_packet);
+    }
+
+    #[test]
+    fn serialize_small_unreliable_packet() {
+        let mut buffer = [0u8; 1300];
+        let packet = Packet::SmallUnreliable {
+            channel_id: 0,
+            messages: vec![vec![0, 0, 0].into(), vec![1, 1, 1].into(), vec![2, 2, 2].into()],
+        };
+
+        let mut b = octets::OctetsMut::with_slice(&mut buffer);
+        packet.to_bytes(&mut b).unwrap();
+
+        let mut b = octets::Octets::with_slice(&buffer);
+        let recv_packet = Packet::from_bytes(&mut b).unwrap();
+        assert_eq!(packet, recv_packet);
+    }
+
+    #[test]
+    fn serialize_reliable_slice_packet() {
+        let mut buffer = [0u8; 1300];
+
+        let packet = Packet::ReliableSlice {
+            packet_sequence: 0,
+            channel_id: 0,
+            slice: Slice {
+                message_id: 0,
+                slice_index: 0,
+                num_slices: 0,
+                payload: vec![5; SLICE_SIZE].into(),
+            },
+        };
+
+        let mut b = octets::OctetsMut::with_slice(&mut buffer);
+        packet.to_bytes(&mut b).unwrap();
+
+        let mut b = octets::Octets::with_slice(&buffer);
+        let recv_packet = Packet::from_bytes(&mut b).unwrap();
+        assert_eq!(packet, recv_packet);
+    }
+
+    #[test]
+    fn serialize_unreliable_slice_packet() {
+        let mut buffer = [0u8; 1300];
+
+        let packet = Packet::UnreliableSlice {
+            channel_id: 0,
+            slice: Slice {
+                message_id: 0,
+                slice_index: 0,
+                num_slices: 0,
+                payload: vec![5; SLICE_SIZE].into(),
+            },
+        };
+
+        let mut b = octets::OctetsMut::with_slice(&mut buffer);
+        packet.to_bytes(&mut b).unwrap();
+
+        let mut b = octets::Octets::with_slice(&buffer);
+        let recv_packet = Packet::from_bytes(&mut b).unwrap();
+        assert_eq!(packet, recv_packet);
+    }
+
+    #[test]
+    fn serialize_ack_packet() {
+        let mut buffer = [0u8; 1300];
+
+        let packet = Packet::Ack {
+            packet_sequence: 0,
+            ack_ranges: vec![3..7, 10..20, 30..100],
+        };
+
+        let mut b = octets::OctetsMut::with_slice(&mut buffer);
+        packet.to_bytes(&mut b).unwrap();
+
+        let mut b = octets::Octets::with_slice(&buffer);
+        let recv_packet = Packet::from_bytes(&mut b).unwrap();
+        assert_eq!(packet, recv_packet);
+    }
+
+    #[test]
+    fn serialize_disconnect_packet() {
+        let mut buffer = [0u8; 1300];
+        let packet = Packet::Disconnect;
+
+        let mut b = octets::OctetsMut::with_slice(&mut buffer);
+        packet.to_bytes(&mut b).unwrap();
+
+        let mut b = octets::Octets::with_slice(&buffer);
+        let recv_packet = Packet::from_bytes(&mut b).unwrap();
+        assert_eq!(packet, recv_packet);
+    }
 }
