+++ rechannel/src/remote_connection.rs
@@ -1,23 +1,15 @@
-use crate::channel::{ChannelConfig, DefaultChannel, ReceiveChannel, SendChannel};
+use crate::channels::reliable::{ReceiveChannelReliable, SendChannelReliable};
+use crate::channels::unreliable::{ReceiveChannelUnreliable, SendChannelUnreliable};
+use crate::channels::{ChannelConfig, SendType};
 use crate::error::{DisconnectionReason, RechannelError};
-use crate::packet::{AckData, Packet, Payload};
-
-use crate::reassembly_fragment::{build_fragments, FragmentConfig, ReassemblyFragment};
-use crate::sequence_buffer::SequenceBuffer;
-use crate::timer::Timer;
-
-use bincode::Options;
+use crate::packet::{Packet, Payload};
 use bytes::Bytes;
 use log::error;
+use octets::OctetsMut;
 
-use std::collections::{HashMap, VecDeque};
+use std::collections::{BTreeMap, HashMap};
+use std::ops::Range;
 use std::time::Duration;
-
-#[derive(Debug, Clone)]
-struct SentPacket {
-    time: Duration,
-    ack: bool,
-}
 
 #[derive(Debug)]
 enum ConnectionState {
@@ -25,107 +17,141 @@ enum ConnectionState {
     Disconnected { reason: DisconnectionReason },
 }
 
+const REQUEST_ACK_TIME: Duration = Duration::from_millis(1000);
+
 #[derive(Debug, Clone)]
 pub struct ConnectionConfig {
-    pub max_packet_size: u64,
-    pub sent_packets_buffer_size: usize,
-    pub received_packets_buffer_size: usize,
-    pub rtt_smoothing_factor: f32,
-    pub packet_loss_smoothing_factor: f32,
-    pub heartbeat_time: Duration,
-    pub fragment_config: FragmentConfig,
     pub send_channels_config: Vec<ChannelConfig>,
     pub receive_channels_config: Vec<ChannelConfig>,
 }
 
-#[derive(Debug)]
-pub struct RemoteConnection {
-    state: ConnectionState,
-    sequence: u16,
-    send_channels: HashMap<u8, Box<dyn SendChannel + Send + Sync + 'static>>,
-    receive_channels: HashMap<u8, Box<dyn ReceiveChannel + Send + Sync + 'static>>,
-    heartbeat_timer: Timer,
-    config: ConnectionConfig,
-    reassembly_buffer: SequenceBuffer<ReassemblyFragment>,
-    sent_buffer: SequenceBuffer<SentPacket>,
-    received_buffer: SequenceBuffer<()>,
-    current_time: Duration,
-    rtt: f32,
-    packet_loss: f32,
-    acks: Vec<u16>,
-}
-
-impl SentPacket {
-    fn new(time: Duration) -> Self {
-        Self { time, ack: false }
-    }
+#[derive(Debug, Clone)]
+enum PacketSent {
+    ReliableMessages {
+        channel_id: u8,
+        message_ids: Vec<u64>,
+    },
+    ReliableSliceMessage {
+        channel_id: u8,
+        message_id: u64,
+        slice_index: usize,
+    },
+    // When an ack packet is acknowledged,
+    // We remove all Ack ranges below the largest_acked sent by it
+    Ack {
+        largest_acked_packet: u64,
+    },
 }
 
 impl Default for ConnectionConfig {
     fn default() -> Self {
         Self {
-            max_packet_size: 16 * 1024,
-            sent_packets_buffer_size: 256,
-            received_packets_buffer_size: 256,
-            rtt_smoothing_factor: 0.01,
-            packet_loss_smoothing_factor: 0.1,
-            heartbeat_time: Duration::from_millis(100),
-            fragment_config: FragmentConfig::default(),
-            send_channels_config: DefaultChannel::config(),
-            receive_channels_config: DefaultChannel::config(),
+            send_channels_config: vec![],
+            receive_channels_config: vec![],
         }
     }
 }
 
+#[derive(Debug)]
+pub struct RemoteConnection {
+    state: ConnectionState,
+    packet_sequence: u64,
+    current_time: Duration,
+    sent_packets: BTreeMap<u64, PacketSent>,
+    pending_acks: Vec<Range<u64>>,
+    send_unreliable_channels: HashMap<u8, SendChannelUnreliable>,
+    receive_unreliable_channels: HashMap<u8, ReceiveChannelUnreliable>,
+    send_reliable_channels: HashMap<u8, SendChannelReliable>,
+    receive_reliable_channels: HashMap<u8, ReceiveChannelReliable>,
+    should_generate_ack: bool,
+    last_ack_received_at: Duration, // rtt: f32,
+                                    // packet_loss: f32,
+                                    // acks: Vec<u16>,
+}
+
 impl RemoteConnection {
     pub fn new(current_time: Duration, config: ConnectionConfig) -> Self {
-        config.fragment_config.assert_can_fragment_packet_with_size(config.max_packet_size);
-
-        let heartbeat_timer = Timer::new(current_time, config.heartbeat_time);
-        let reassembly_buffer = SequenceBuffer::with_capacity(config.fragment_config.reassembly_buffer_size);
-        let sent_buffer = SequenceBuffer::with_capacity(config.sent_packets_buffer_size);
-        let received_buffer = SequenceBuffer::with_capacity(config.received_packets_buffer_size);
-
-        let mut send_channels = HashMap::new();
+        let mut send_unreliable_channels = HashMap::new();
+        let mut send_reliable_channels = HashMap::new();
         for channel_config in config.send_channels_config.iter() {
-            let (send_channel, _) = channel_config.new_channels();
-            let channel_id = channel_config.channel_id();
-            let old_channel = send_channels.insert(channel_id, send_channel);
-            assert!(old_channel.is_none(), "already exists send channel with id {}", channel_id);
+            match channel_config.send_type {
+                SendType::Unreliable => {
+                    let channel = SendChannelUnreliable::new(channel_config.channel_id, channel_config.max_memory_usage_bytes);
+                    let old_channel = send_unreliable_channels.insert(channel_config.channel_id, channel);
+                    assert!(
+                        old_channel.is_none(),
+                        "already exists send channel with id {}",
+                        channel_config.channel_id
+                    );
+                }
+                SendType::ReliableOrdered { resend_time } | SendType::ReliableUnordered { resend_time } => {
+                    let channel = SendChannelReliable::new(channel_config.channel_id, resend_time, channel_config.max_memory_usage_bytes);
+                    let old_channel = send_reliable_channels.insert(channel_config.channel_id, channel);
+                    assert!(
+                        old_channel.is_none(),
+                        "already exists send channel with id {}",
+                        channel_config.channel_id
+                    );
+                }
+            }
         }
 
-        let mut receive_channels = HashMap::new();
+        let mut receive_unreliable_channels = HashMap::new();
+        let mut receive_reliable_channels = HashMap::new();
         for channel_config in config.receive_channels_config.iter() {
-            let (_, receive_channel) = channel_config.new_channels();
-            let channel_id = channel_config.channel_id();
-            let old_channel = receive_channels.insert(channel_id, receive_channel);
-            assert!(old_channel.is_none(), "already exists receive channel with id {}", channel_id);
+            match channel_config.send_type {
+                SendType::Unreliable => {
+                    let channel = ReceiveChannelUnreliable::new(channel_config.channel_id, channel_config.max_memory_usage_bytes);
+                    let old_channel = receive_unreliable_channels.insert(channel_config.channel_id, channel);
+                    assert!(
+                        old_channel.is_none(),
+                        "already exists receive channel with id {}",
+                        channel_config.channel_id
+                    );
+                }
+                SendType::ReliableOrdered { .. } => {
+                    let channel = ReceiveChannelReliable::new(channel_config.channel_id, channel_config.max_memory_usage_bytes, true);
+                    let old_channel = receive_reliable_channels.insert(channel_config.channel_id, channel);
+                    assert!(
+                        old_channel.is_none(),
+                        "already exists receive channel with id {}",
+                        channel_config.channel_id
+                    );
+                }
+                SendType::ReliableUnordered { .. } => {
+                    let channel = ReceiveChannelReliable::new(channel_config.channel_id, channel_config.max_memory_usage_bytes, false);
+                    let old_channel = receive_reliable_channels.insert(channel_config.channel_id, channel);
+                    assert!(
+                        old_channel.is_none(),
+                        "already exists receive channel with id {}",
+                        channel_config.channel_id
+                    );
+                }
+            }
         }
 
         Self {
             state: ConnectionState::Connected,
-            send_channels,
-            receive_channels,
-            heartbeat_timer,
-            sequence: 0,
-            reassembly_buffer,
-            sent_buffer,
-            received_buffer,
+            packet_sequence: 0,
             current_time,
-            config,
-            rtt: 0.0,
-            packet_loss: 0.0,
-            acks: vec![],
+            sent_packets: BTreeMap::new(),
+            pending_acks: Vec::new(),
+            send_unreliable_channels,
+            receive_unreliable_channels,
+            send_reliable_channels,
+            receive_reliable_channels,
+            should_generate_ack: false,
+            last_ack_received_at: current_time,
         }
     }
 
-    pub fn rtt(&self) -> f32 {
-        self.rtt
-    }
+    // pub fn rtt(&self) -> f32 {
+    //     self.rtt
+    // }
 
-    pub fn packet_loss(&self) -> f32 {
-        self.packet_loss
-    }
+    // pub fn packet_loss(&self) -> f32 {
+    //     self.packet_loss
+    // }
 
     pub fn is_connected(&self) -> bool {
         matches!(self.state, ConnectionState::Connected)
@@ -149,19 +175,31 @@ impl RemoteConnection {
         };
     }
 
-    pub fn can_send_message<I: Into<u8>>(&self, channel_id: I) -> bool {
-        let channel = self.send_channels.get(&channel_id.into()).expect("invalid channel id");
-        channel.can_send_message()
-    }
+    // pub fn can_send_message<I: Into<u8>>(&self, channel_id: I) -> bool {
+    //     let channel = self.send_channels.get(&channel_id.into()).expect("invalid channel id");
+    //     channel.can_send_message()
+    // }
 
     pub fn send_message<I: Into<u8>, B: Into<Bytes>>(&mut self, channel_id: I, message: B) {
-        let channel = self.send_channels.get_mut(&channel_id.into()).expect("invalid channel id");
-        channel.send_message(message.into(), self.current_time);
+        let channel_id = channel_id.into();
+        if let Some(reliable_channel) = self.send_reliable_channels.get_mut(&channel_id) {
+            reliable_channel.send_message(message.into());
+        } else if let Some(unreliable_channel) = self.send_unreliable_channels.get_mut(&channel_id) {
+            unreliable_channel.send_message(message.into());
+        } else {
+            panic!("Tried to send message to invalid channel {channel_id}");
+        }
     }
 
-    pub fn receive_message<I: Into<u8>>(&mut self, channel_id: I) -> Option<Payload> {
-        let channel = self.receive_channels.get_mut(&channel_id.into()).expect("invalid channel id");
-        channel.receive_message()
+    pub fn receive_message<I: Into<u8>>(&mut self, channel_id: I) -> Option<Bytes> {
+        let channel_id = channel_id.into();
+        if let Some(reliable_channel) = self.receive_reliable_channels.get_mut(&channel_id) {
+            reliable_channel.receive_message()
+        } else if let Some(unreliable_channel) = self.receive_unreliable_channels.get_mut(&channel_id) {
+            unreliable_channel.receive_message()
+        } else {
+            panic!("Tried to receive message from invalid channel {channel_id}");
+        }
     }
 
     pub fn advance_time(&mut self, duration: Duration) {
@@ -169,272 +207,337 @@ impl RemoteConnection {
     }
 
     pub fn update(&mut self) -> Result<(), RechannelError> {
-        if let Some(reason) = self.disconnected() {
-            return Err(RechannelError::ClientDisconnected(reason));
-        }
+        //     if let Some(reason) = self.disconnected() {
+        //         return Err(RechannelError::ClientDisconnected(reason));
+        //     }
 
-        for (&channel_id, send_channel) in self.send_channels.iter() {
-            if let Some(error) = send_channel.error() {
-                let reason = DisconnectionReason::SendChannelError { channel_id, error };
-                self.state = ConnectionState::Disconnected { reason };
-                return Err(RechannelError::ClientDisconnected(reason));
-            }
-        }
+        //     for (&channel_id, send_channel) in self.send_channels.iter() {
+        //         if let Some(error) = send_channel.error() {
+        //             let reason = DisconnectionReason::SendChannelError { channel_id, error };
+        //             self.state = ConnectionState::Disconnected { reason };
+        //             return Err(RechannelError::ClientDisconnected(reason));
+        //         }
+        //     }
 
-        for (&channel_id, receive_channel) in self.receive_channels.iter() {
-            if let Some(error) = receive_channel.error() {
-                let reason = DisconnectionReason::ReceiveChannelError { channel_id, error };
-                self.state = ConnectionState::Disconnected { reason };
-                return Err(RechannelError::ClientDisconnected(reason));
-            }
-        }
-
-        for ack in self.acks.drain(..) {
-            for channel in self.send_channels.values_mut() {
-                channel.process_ack(ack);
-            }
-        }
-
-        self.update_packet_loss();
+        //     for (&channel_id, receive_channel) in self.receive_channels.iter() {
+        //         if let Some(error) = receive_channel.error() {
+        //             let reason = DisconnectionReason::ReceiveChannelError { channel_id, error };
+        //             self.state = ConnectionState::Disconnected { reason };
+        //             return Err(RechannelError::ClientDisconnected(reason));
+        //         }
+        //     }
 
         Ok(())
     }
 
     pub fn process_packet(&mut self, packet: &[u8]) -> Result<(), RechannelError> {
-        if let Some(reason) = self.disconnected() {
-            return Err(RechannelError::ClientDisconnected(reason));
+        let mut octets = octets::Octets::with_slice(packet);
+        let Ok(packet) = Packet::from_bytes(&mut octets) else {
+            return Err(RechannelError::Serialization);
+        };
+
+        if packet.is_ack_eliciting() {
+            self.should_generate_ack = true;
         }
 
-        let packet: Packet = bincode::options().deserialize(packet)?;
-
-        let channels_packet_data = match packet {
-            Packet::Normal {
-                sequence,
-                ack_data,
-                channels_packet_data,
+        match packet {
+            Packet::SmallReliable {
+                packet_sequence,
+                channel_id,
+                messages,
             } => {
-                self.received_buffer.insert(sequence, ());
-                self.update_acket_packets(ack_data.ack, ack_data.ack_bits);
-                channels_packet_data
+                self.add_pending_ack(packet_sequence);
+                let Some(channel) = self.receive_reliable_channels.get_mut(&channel_id) else {
+                    // TODO: self.error = channel not found;
+                    return Err(RechannelError::ChannelNotFound(channel_id));
+                };
+
+                for (message_id, message) in messages {
+                    channel.process_message(message, message_id);
+                }
             }
-            Packet::Fragment {
-                sequence,
-                ack_data,
-                fragment_data,
-            } => {
-                self.update_acket_packets(ack_data.ack, ack_data.ack_bits);
+            Packet::SmallUnreliable { channel_id, messages } => {
+                let Some(channel) = self.receive_unreliable_channels.get_mut(&channel_id) else {
+                    // TODO: self.error = channel not found;
+                    return Err(RechannelError::ChannelNotFound(channel_id));
+                };
 
-                let packet = self.reassembly_buffer.handle_fragment(
-                    sequence,
-                    fragment_data,
-                    self.config.max_packet_size,
-                    &self.config.fragment_config,
-                )?;
-                match packet {
-                    None => return Ok(()),
-                    Some(packet) => {
-                        // Only consider the packet received when the fragment is completed
-                        self.received_buffer.insert(sequence, ());
-                        packet
+                for message in messages {
+                    channel.process_message(message);
+                }
+            }
+            Packet::ReliableSlice {
+                packet_sequence,
+                channel_id,
+                slice,
+            } => {
+                self.add_pending_ack(packet_sequence);
+                let Some(channel) = self.receive_reliable_channels.get_mut(&channel_id) else {
+                    // TODO: self.error = channel not found;
+                    return Err(RechannelError::ChannelNotFound(channel_id));
+                };
+
+                channel.process_slice(slice);
+            }
+            Packet::UnreliableSlice { channel_id, slice } => {
+                let Some(channel) = self.receive_unreliable_channels.get_mut(&channel_id) else {
+                    // TODO: self.error = channel not found;
+                    return Err(RechannelError::ChannelNotFound(channel_id));
+                };
+
+                channel.process_slice(slice, self.current_time);
+            }
+            Packet::Ack {
+                packet_sequence,
+                ack_ranges,
+            } => {
+                self.last_ack_received_at = self.current_time;
+                self.add_pending_ack(packet_sequence);
+
+                // Create list with just new acks
+                // This prevents DoS from huge ack ranges
+                let mut new_acks: Vec<u64> = Vec::new();
+                for range in ack_ranges {
+                    for (&sequence, _) in self.sent_packets.range(range) {
+                        new_acks.push(sequence)
+                    }
+                }
+
+                for packet_sequence in new_acks {
+                    let sent_packet = self.sent_packets.remove(&packet_sequence).unwrap();
+
+                    match sent_packet {
+                        PacketSent::ReliableMessages { channel_id, message_ids } => {
+                            let reliable_channel = self.send_reliable_channels.get_mut(&channel_id).unwrap();
+                            for message_id in message_ids {
+                                reliable_channel.process_message_ack(message_id);
+                            }
+                        }
+                        PacketSent::ReliableSliceMessage {
+                            channel_id,
+                            message_id,
+                            slice_index,
+                        } => {
+                            let reliable_channel = self.send_reliable_channels.get_mut(&channel_id).unwrap();
+                            reliable_channel.process_slice_message_ack(message_id, slice_index);
+                        }
+                        PacketSent::Ack { largest_acked_packet } => {
+                            self.acked_largest(largest_acked_packet);
+                        }
                     }
                 }
             }
-            Packet::Heartbeat { ack_data } => {
-                self.update_acket_packets(ack_data.ack, ack_data.ack_bits);
-                return Ok(());
-            }
-            Packet::Disconnect { reason } => {
-                self.state = ConnectionState::Disconnected { reason };
-                return Ok(());
-            }
-        };
-
-        for channel_packet_data in channels_packet_data.into_iter() {
-            let receive_channel = match self.receive_channels.get_mut(&channel_packet_data.channel_id) {
-                Some(c) => c,
-                None => {
-                    let reason = DisconnectionReason::InvalidChannelId(channel_packet_data.channel_id);
-                    self.state = ConnectionState::Disconnected { reason };
-                    return Err(RechannelError::ClientDisconnected(reason));
-                }
-            };
-
-            receive_channel.process_messages(channel_packet_data.messages);
+            Packet::RequestAck => {}
+            Packet::Disconnect => {}
         }
 
         Ok(())
     }
 
     pub fn get_packets_to_send(&mut self) -> Result<Vec<Payload>, RechannelError> {
-        if let Some(reason) = self.disconnected() {
-            return Err(RechannelError::ClientDisconnected(reason));
+        let mut packets: Vec<Packet> = vec![];
+
+        for channel in self.send_unreliable_channels.values_mut() {
+            packets.append(&mut channel.get_messages_to_send());
         }
 
-        let sequence = self.sequence;
-        // Aproximated header size for the packet
-        const HEADER_SIZE: u64 = 20;
-        let mut available_bytes = self.config.max_packet_size - HEADER_SIZE;
-        let mut channels_packet_data = vec![];
-        for send_channel in self.send_channels.values_mut() {
-            if let Some(channel_packet_data) = send_channel.get_messages_to_send(available_bytes, sequence, self.current_time) {
-                available_bytes -= bincode::options().serialized_size(&channel_packet_data)?;
-                channels_packet_data.push(channel_packet_data);
+        for channel in self.send_reliable_channels.values_mut() {
+            packets.append(&mut channel.get_messages_to_send(&mut self.packet_sequence, self.current_time));
+        }
+
+        if !self.pending_acks.is_empty() && self.should_generate_ack {
+            let ack_packet = Packet::Ack {
+                packet_sequence: self.packet_sequence,
+                ack_ranges: self.pending_acks.clone(),
+            };
+            self.packet_sequence += 1;
+            self.should_generate_ack = false;
+            packets.push(ack_packet);
+        }
+
+        for packet in packets.iter() {
+            match packet {
+                Packet::SmallReliable {
+                    packet_sequence,
+                    channel_id,
+                    messages,
+                } => {
+                    self.sent_packets.insert(
+                        *packet_sequence,
+                        PacketSent::ReliableMessages {
+                            channel_id: *channel_id,
+                            message_ids: messages.iter().map(|(id, _)| *id).collect(),
+                        },
+                    );
+                }
+                Packet::ReliableSlice {
+                    packet_sequence,
+                    channel_id,
+                    slice,
+                } => {
+                    self.sent_packets.insert(
+                        *packet_sequence,
+                        PacketSent::ReliableSliceMessage {
+                            channel_id: *channel_id,
+                            message_id: slice.message_id,
+                            slice_index: slice.slice_index,
+                        },
+                    );
+                }
+                Packet::Ack {
+                    packet_sequence,
+                    ack_ranges,
+                } => {
+                    let last_range = ack_ranges.last().unwrap();
+                    let largest_acked_packet = last_range.end - 1;
+                    self.sent_packets.insert(*packet_sequence, PacketSent::Ack { largest_acked_packet });
+                }
+                _ => {}
             }
         }
 
-        if !channels_packet_data.is_empty() {
-            // self.sequence = self.sequence.wrapping_add(1);
-            // let packet_size = bincode::options().serialized_size(&channels_packet_data)?;
-            // let ack_data = self.received_buffer.ack_data(self.received_buffer.sequence().wrapping_sub(1));
-
-            // let sent_packet = SentPacket::new(self.current_time);
-            // self.sent_buffer.insert(sequence, sent_packet);
-
-            // let packets: Vec<Payload> = if packet_size > self.config.fragment_config.fragment_above {
-            //     build_fragments(channels_packet_data, sequence, ack_data, &self.config.fragment_config)?
-            // } else {
-            //     let packet = Packet::Normal {
-            //         sequence,
-            //         ack_data,
-            //         channels_packet_data,
-            //     };
-            //     let packet = bincode::options().serialize(&packet)?;
-            //     vec![packet]
-            // };
-
-            // self.heartbeat_timer.reset(self.current_time);
-            // return Ok(packets);
-        } else if self.heartbeat_timer.is_finished(self.current_time) {
-            // let ack_data = self.received_buffer.ack_data(self.received_buffer.sequence().wrapping_sub(1));
-            // let packet = Packet::Heartbeat { ack_data };
-            // let packet = bincode::options().serialize(&packet)?;
-
-            // self.heartbeat_timer.reset(self.current_time);
-            // return Ok(vec![packet]);
+        let has_ack_eliciting_packet = packets.iter().any(Packet::is_ack_eliciting);
+        if !has_ack_eliciting_packet && !self.sent_packets.is_empty() {
+            if self.last_ack_received_at + REQUEST_ACK_TIME < self.current_time {
+                packets.push(Packet::RequestAck);
+            }
         }
 
-        // TODO: should we return Option<Vec>?
-        Ok(vec![])
+        let mut buffer = [0u8; 1400];
+        let mut serialized_packets = Vec::with_capacity(packets.len());
+        for packet in packets {
+            let mut oct = OctetsMut::with_slice(&mut buffer);
+            let len = packet.to_bytes(&mut oct).map_err(|_| RechannelError::Serialization)?;
+            serialized_packets.push(buffer[..len].to_vec());
+        }
+
+        Ok(serialized_packets)
     }
 
-    fn update_acket_packets(&mut self, ack: u16, mut ack_bits: u32) {
-        for i in 0..32 {
-            if ack_bits & 1 != 0 {
-                let ack_sequence = ack.wrapping_sub(i);
-                if let Some(ref mut sent_packet) = self.sent_buffer.get_mut(ack_sequence) {
-                    if !sent_packet.ack {
-                        self.acks.push(ack_sequence);
-                        sent_packet.ack = true;
-
-                        // Update RTT
-                        let rtt = (self.current_time - sent_packet.time).as_secs_f32() * 1000.;
-
-                        if self.rtt == 0.0 || self.rtt < f32::EPSILON {
-                            self.rtt = rtt;
-                        } else {
-                            self.rtt += (rtt - self.rtt) * self.config.rtt_smoothing_factor;
-                        }
-                    }
-                }
-            }
-            ack_bits >>= 1;
-        }
-    }
-
-    fn update_packet_loss(&mut self) {
-        let sample_size = self.config.sent_packets_buffer_size;
-        let base_sequence = self.sent_buffer.sequence().wrapping_sub(sample_size as u16);
-
-        let mut packets_dropped = 0;
-        let mut packets_sent = 0;
-        for i in 0..sample_size {
-            if let Some(sent_packet) = self.sent_buffer.get(base_sequence.wrapping_add(i as u16)) {
-                packets_sent += 1;
-                let secs_since_sent = (self.current_time - sent_packet.time).as_secs_f32();
-                if !sent_packet.ack && secs_since_sent > self.rtt * 1.5 {
-                    packets_dropped += 1;
-                }
-            }
-        }
-
-        if packets_sent == 0 {
+    fn add_pending_ack(&mut self, sequence: u64) {
+        if self.pending_acks.is_empty() {
+            self.pending_acks.push(sequence..sequence + 1);
             return;
         }
 
-        let packet_loss = packets_dropped as f32 / packets_sent as f32;
-        if self.packet_loss == 0.0 || self.packet_loss < f32::EPSILON {
-            self.packet_loss = packet_loss;
-        } else {
-            self.packet_loss += (packet_loss - self.packet_loss) * self.config.packet_loss_smoothing_factor;
+        for index in 0..self.pending_acks.len() {
+            let range = &mut self.pending_acks[index];
+            if range.contains(&sequence) {
+                // Sequence already contained in this range
+                return;
+            }
+
+            if range.start == sequence + 1 {
+                // New sequence is just before this range
+                range.start = sequence;
+                return;
+            } else if range.end == sequence {
+                // New sequence is just after this range
+                range.end = sequence + 1;
+
+                // Check if we can merge with the range just after it
+                let next_index = index + 1;
+                if next_index < self.pending_acks.len() && self.pending_acks[index].end == self.pending_acks[next_index].start {
+                    self.pending_acks[index].end = self.pending_acks[next_index].end;
+                    self.pending_acks.remove(next_index);
+                }
+
+                return;
+            } else if self.pending_acks[index].start > sequence + 1 {
+                // New sequence is before this range and not extensible to it
+                // Add new range to the left
+                self.pending_acks.insert(index, sequence..sequence + 1);
+                return;
+            }
+        }
+
+        // New sequence was not before or adjacent to any range
+        // Add new range with only this sequence at the end
+        self.pending_acks.push(sequence..sequence + 1);
+    }
+
+    fn acked_largest(&mut self, largest_ack: u64) {
+        while self.pending_acks.len() != 0 {
+            let range = &mut self.pending_acks[0];
+
+            // Largest ack is below the range, stop checking
+            if largest_ack < range.start {
+                return;
+            }
+
+            // Largest ack is above the range, remove it
+            if range.end <= largest_ack {
+                self.pending_acks.remove(0);
+                continue;
+            }
+
+            // Largest ack is contained in the range
+            // Update start
+            range.start = largest_ack + 1;
+            if range.is_empty() {
+                self.pending_acks.remove(0);
+            }
+
+            return;
         }
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::packet::AckData;
-
     use super::*;
 
     #[test]
-    fn round_time_trip() {
+    fn pending_acks() {
         let mut connection = RemoteConnection::new(Duration::ZERO, ConnectionConfig::default());
+        connection.add_pending_ack(3);
+        assert_eq!(connection.pending_acks, vec![3..4]);
 
-        let message: Bytes = vec![1, 2, 3].into();
-        let mut ack_data = AckData { ack: 0, ack_bits: 1 };
-        for _ in 0..16 {
-            connection.send_message(1, message.clone());
-            assert!(!connection.get_packets_to_send().unwrap().is_empty());
+        connection.add_pending_ack(4);
+        assert_eq!(connection.pending_acks, vec![3..5]);
 
-            connection.advance_time(Duration::from_millis(100));
-            connection.update_acket_packets(ack_data.ack, ack_data.ack_bits);
+        connection.add_pending_ack(2);
+        assert_eq!(connection.pending_acks, vec![2..5]);
 
-            ack_data.ack += 1;
-        }
+        connection.add_pending_ack(0);
+        assert_eq!(connection.pending_acks, vec![0..1, 2..5]);
 
-        assert_eq!(connection.rtt(), 100.);
+        connection.add_pending_ack(7);
+        assert_eq!(connection.pending_acks, vec![0..1, 2..5, 7..8]);
+
+        connection.add_pending_ack(1);
+        assert_eq!(connection.pending_acks, vec![0..5, 7..8]);
+
+        connection.add_pending_ack(5);
+        assert_eq!(connection.pending_acks, vec![0..6, 7..8]);
+
+        connection.add_pending_ack(6);
+        assert_eq!(connection.pending_acks, vec![0..8]);
     }
 
     #[test]
-    fn packet_loss() {
+    fn ack_pending_acks() {
         let mut connection = RemoteConnection::new(Duration::ZERO, ConnectionConfig::default());
-
-        let message: Bytes = vec![1, 2, 3].into();
-        let mut ack_data = AckData { ack: 0, ack_bits: 1 };
-        for i in 0..32 {
-            connection.send_message(1, message.clone());
-            assert!(!connection.get_packets_to_send().unwrap().is_empty());
-
-            // 50% packet loss
-            if i % 2 == 0 {
-                connection.update_acket_packets(ack_data.ack, ack_data.ack_bits);
-            }
-            connection.advance_time(Duration::from_millis(100));
-
-            ack_data.ack += 1;
+        for i in 0..10 {
+            connection.add_pending_ack(i);
         }
 
-        connection.update_packet_loss();
-        assert_eq!(connection.packet_loss(), 0.5);
-    }
+        assert_eq!(connection.pending_acks, vec![0..10]);
 
-    #[test]
-    fn confirm_only_completed_fragmented_packet() {
-        let config = ConnectionConfig::default();
-        let mut connection = RemoteConnection::new(Duration::ZERO, config);
-        let message = vec![7u8; 2500];
-        connection.send_message(0, message.clone());
+        connection.acked_largest(0);
+        assert_eq!(connection.pending_acks, vec![1..10]);
 
-        let packets = connection.get_packets_to_send().unwrap();
-        assert!(packets.len() > 1);
-        for packet in packets.iter() {
-            assert!(!connection.received_buffer.exists(0));
-            connection.process_packet(packet).unwrap();
-        }
-        // After all fragments are received it should be considered received
-        assert!(connection.received_buffer.exists(0));
+        connection.acked_largest(3);
+        assert_eq!(connection.pending_acks, vec![4..10]);
 
-        let received_message = connection.receive_message(0).unwrap();
-        assert_eq!(message, received_message);
+        connection.add_pending_ack(0);
+        assert_eq!(connection.pending_acks, vec![0..1, 4..10]);
+        connection.acked_largest(5);
+        assert_eq!(connection.pending_acks, vec![6..10]);
+
+        connection.add_pending_ack(0);
+        assert_eq!(connection.pending_acks, vec![0..1, 6..10]);
+        connection.acked_largest(10);
+        assert_eq!(connection.pending_acks, vec![]);
     }
 }
