+++ renet/src/server.rs
@@ -11,7 +11,7 @@ use std::{
 };
 
 use log::error;
-use rechannel::{disconnect_packet, error::DisconnectionReason, server::RechannelServer, Bytes};
+use rechannel::{error::DisconnectionReason, server::RechannelServer, Bytes};
 use renetcode::{NetcodeServer, ServerResult, NETCODE_KEY_BYTES, NETCODE_USER_DATA_BYTES};
 
 /// A server that can establish authenticated connections with multiple clients.
@@ -20,7 +20,7 @@ use renetcode::{NetcodeServer, ServerResult, NETCODE_KEY_BYTES, NETCODE_USER_DAT
 #[cfg_attr(feature = "bevy", derive(bevy_ecs::system::Resource))]
 pub struct RenetServer {
     socket: UdpSocket,
-    reliable_server: RechannelServer<u64>,
+    reliable_server: RechannelServer,
     netcode_server: NetcodeServer,
     bandwidth_smoothing_factor: f32,
     clients_packet_info: HashMap<SocketAddr, ClientPacketInfo>,
@@ -82,8 +82,8 @@ impl RenetServer {
         connection_config: RenetConnectionConfig,
         socket: UdpSocket,
     ) -> Result<Self, std::io::Error> {
-        let buffer = vec![0u8; connection_config.max_packet_size as usize].into_boxed_slice();
-        let bandwidth_smoothing_factor = connection_config.bandwidth_smoothing_factor;
+        let buffer = vec![0u8; 2000].into_boxed_slice();
+        let bandwidth_smoothing_factor = 0.05;
         let reliable_server = RechannelServer::new(current_time, connection_config.to_connection_config());
 
         // For unsecure connections we use an fixed private key.
@@ -153,26 +153,26 @@ impl RenetServer {
     }
 
     /// Returns the client's network info if the client exits.
-    pub fn network_info(&self, client_id: u64) -> Option<NetworkInfo> {
-        let addr = match self.netcode_server.client_addr(client_id) {
-            Some(addr) => addr,
-            None => return None,
-        };
+    // pub fn network_info(&self, client_id: u64) -> Option<NetworkInfo> {
+    //     let addr = match self.netcode_server.client_addr(client_id) {
+    //         Some(addr) => addr,
+    //         None => return None,
+    //     };
 
-        let client_packet_info = self.clients_packet_info.get(&addr).unwrap();
+    //     let client_packet_info = self.clients_packet_info.get(&addr).unwrap();
 
-        let sent_kbps = client_packet_info.sent_kbps;
-        let received_kbps = client_packet_info.received_kbps;
-        let rtt = self.reliable_server.client_rtt(client_id);
-        let packet_loss = self.reliable_server.client_packet_loss(client_id);
+    //     let sent_kbps = client_packet_info.sent_kbps;
+    //     let received_kbps = client_packet_info.received_kbps;
+    //     let rtt = self.reliable_server.client_rtt(client_id);
+    //     let packet_loss = self.reliable_server.client_packet_loss(client_id);
 
-        Some(NetworkInfo {
-            received_kbps,
-            sent_kbps,
-            rtt,
-            packet_loss,
-        })
-    }
+    //     Some(NetworkInfo {
+    //         received_kbps,
+    //         sent_kbps,
+    //         rtt,
+    //         packet_loss,
+    //     })
+    // }
 
     /// Advances the server by duration, and receive packets from the network.
     pub fn update(&mut self, duration: Duration) -> Result<(), io::Error> {
@@ -221,20 +221,7 @@ impl RenetServer {
         // Handle disconnected clients from Rechannel
         while let Some((client_id, reason)) = self.reliable_server.disconnected_client() {
             self.events.push_back(ServerEvent::ClientDisconnected(client_id));
-            if reason != DisconnectionReason::DisconnectedByClient {
-                match disconnect_packet(reason) {
-                    Err(e) => error!("Failed to serialize disconnect packet: {}", e),
-                    Ok(packet) => match self.netcode_server.generate_payload_packet(client_id, &packet) {
-                        Err(e) => error!("Failed to encrypt disconnect packet: {}", e),
-                        Ok((addr, payload)) => {
-                            for _ in 0..NUM_DISCONNECT_PACKETS_TO_SEND {
-                                self.socket.send_to(payload, addr)?;
-                            }
-                        }
-                    },
-                }
-            }
-            self.netcode_server.disconnect(client_id);
+            self.disconnect(client_id);
         }
 
         for packet_info in self.clients_packet_info.values_mut() {
@@ -245,23 +232,23 @@ impl RenetServer {
     }
 
     /// Receive a message from a client over a channel.
-    pub fn receive_message<I: Into<u8>>(&mut self, client_id: u64, channel_id: I) -> Option<Vec<u8>> {
-        self.reliable_server.receive_message(&client_id, channel_id)
+    pub fn receive_message<I: Into<u8>>(&mut self, client_id: u64, channel_id: I) -> Option<Bytes> {
+        self.reliable_server.receive_message(client_id, channel_id)
     }
 
     /// Verifies if a message can be sent to a client over a channel.
-    pub fn can_send_message<I: Into<u8>>(&self, client_id: u64, channel_id: I) -> bool {
-        self.reliable_server.can_send_message(&client_id, channel_id)
-    }
+    // pub fn can_send_message<I: Into<u8>>(&self, client_id: u64, channel_id: I) -> bool {
+    //     self.reliable_server.can_send_message(&client_id, channel_id)
+    // }
 
     /// Send a message to a client over a channel.
     pub fn send_message<I: Into<u8>, B: Into<Bytes>>(&mut self, client_id: u64, channel_id: I, message: B) {
-        self.reliable_server.send_message(&client_id, channel_id, message);
+        self.reliable_server.send_message(client_id, channel_id, message);
     }
 
     /// Send a message to all client, except the specified one, over a channel.
     pub fn broadcast_message_except<I: Into<u8>, B: Into<Bytes>>(&mut self, client_id: u64, channel_id: I, message: B) {
-        self.reliable_server.broadcast_message_except(&client_id, channel_id, message)
+        self.reliable_server.broadcast_message_except(client_id, channel_id, message)
     }
 
     /// Send a message to all client over a channel.
@@ -272,7 +259,7 @@ impl RenetServer {
     /// Send packets to connected clients.
     pub fn send_packets(&mut self) -> Result<(), io::Error> {
         for client_id in self.reliable_server.connections_id().into_iter() {
-            let packets = match self.reliable_server.get_packets_to_send(&client_id) {
+            let packets = match self.reliable_server.get_packets_to_send(client_id) {
                 Ok(p) => p,
                 Err(e) => {
                     error!("Failed to get packets from {}: {}", client_id, e);
@@ -329,7 +316,7 @@ fn handle_server_result(
     current_time: Duration,
     bandwidth_smoothing_factor: f32,
     socket: &UdpSocket,
-    reliable_server: &mut RechannelServer<u64>,
+    reliable_server: &mut RechannelServer,
     packet_infos: &mut HashMap<SocketAddr, ClientPacketInfo>,
     events: &mut VecDeque<ServerEvent>,
 ) -> Result<(), io::Error> {
@@ -339,11 +326,11 @@ fn handle_server_result(
             send_to(current_time, socket, packet_infos, payload, addr)?;
         }
         ServerResult::Payload { client_id, payload } => {
-            if !reliable_server.is_connected(&client_id) {
-                reliable_server.add_connection(&client_id);
+            if !reliable_server.is_connected(client_id) {
+                reliable_server.add_connection(client_id);
             }
 
-            if let Err(e) = reliable_server.process_packet_from(payload, &client_id) {
+            if let Err(e) = reliable_server.process_packet_from(payload, client_id) {
                 log::error!("Error while processing payload for {}: {}", client_id, e)
             }
         }
@@ -353,14 +340,14 @@ fn handle_server_result(
             addr,
             payload,
         } => {
-            reliable_server.add_connection(&client_id);
+            reliable_server.add_connection(client_id);
             packet_infos.insert(addr, ClientPacketInfo::new(bandwidth_smoothing_factor));
             events.push_back(ServerEvent::ClientConnected(client_id, user_data));
             send_to(current_time, socket, packet_infos, payload, addr)?;
         }
         ServerResult::ClientDisconnected { client_id, addr, payload } => {
             events.push_back(ServerEvent::ClientDisconnected(client_id));
-            reliable_server.remove_connection(&client_id);
+            reliable_server.remove_connection(client_id);
             packet_infos.remove(&addr);
             if let Some(payload) = payload {
                 for _ in 0..NUM_DISCONNECT_PACKETS_TO_SEND {
